+++ tests/__helpers__/helpers.ts
@@ -1,0 +1,7 @@
+import * as fs from 'fs-jetpack'
+
+export function assertBuildPresent() {
+  if (fs.exists('../../dist')) {
+    throw new Error(`Please build package before running this test`)
+  }
+}
