+++ tests/unit/graphqlSchema/relation1ToN.test.ts
@@ -1,5 +1,5 @@
 import { Specs } from '../../specs'
-import { testGraphqlSchema } from '../../__helpers__'
+import { testGraphqlSchema } from '../../__helpers__/testers'
 
 testGraphqlSchema(Specs.relation1ToNReverse)
 testGraphqlSchema(Specs.relation1ToNReverseAndOptional)
