+++ tests/unit/typescriptDeclarationFile/enumDoc.test.ts
@@ -1,5 +1,5 @@
 import endent from 'endent'
-import { testGeneratedModules } from '../../__helpers__'
+import { testGeneratedModules } from '../../__helpers__/testers'
 
 testGeneratedModules({
   description:
